Moved types into their own files
