+++ MyMoney.Web/ClientApp/src/app/pages/transactions/import/field.enum.ts
@@ -1,0 +1,7 @@
+export enum Field {
+   Description = 'Description',
+   Notes = 'Notes',
+   Amount = 'Amount',
+   Date = 'Date',
+   Ignore = 'Ignore'
+}
